+++ src/util.ts
@@ -9,8 +9,10 @@ export function resolveHostFromRequest(req: IncomingMessage|Http2ServerRequest):
   ;
 }
 
-export function resolveIpFromRequest(req: IncomingMessage|Http2ServerRequest): string {
-  const forwardedFor: string = req.headers['x-forwarded-for'] as string|undefined;
+export function resolveIpFromRequest(req: IncomingMessage|Http2ServerRequest, behindProxy = false): string {
+  // for security reasons, we should never ASSUME the server is behind a proxy
+  // and only support `x-forwarded-for` is explicitly enabled
+  const forwardedFor: string = behindProxy && req.headers['x-forwarded-for'] as string|undefined;
   return forwardedFor ||
     req?.socket?.remoteAddress ||
     'unknown'